+++ minifica.data/ViewModels/BaseViewModel.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace minifica.data.ViewModels
+{
+	public class BaseViewModel
+	{
+		public int Id { get; set; }
+		public bool Active { get; set; }
+	}
+}