+++ minifica.api/CHANGELOG.md
@@ -1,4 +1,7 @@
 ﻿### Changelog
 
 ## [DevOps-8] - [2023-06-02] - Albino Martinez
+* Added BaseViewModel class.
+
+## [DevOps-8] - [2023-06-02] - Albino Martinez
 * Added changelog file, to records commits.